requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
httpx
motor
requests
pandas
sentence-transformers 
faiss-cpu